Saving data into h5
I have a lot of problems trying to use the code from DnCNN-keras in my database.
The data from cygnus are too large, so I had to save the arrays recursively in
a h5 file. Now seems work well but I need to check if all images are being pass
to algorithm